refactor: organize client features
