+++ apps/client/src/features/auth/services/auth.js
@@ -1,4 +1,4 @@
-import http from './http.js';
+import http from '../../../shared/services/http.js';
 
 export const login = (agentId, workerSid, identity) =>
   http.post('/auth/login', { agentId, workerSid, identity }).then((r) => r.data);
