+++ apps/client/src/features/softphone/components/PopoutButton.jsx
