+++ apps/client/src/features/tasks/components/Presence.jsx
@@ -3,7 +3,7 @@ import { useState, useEffect, useMemo } from 'react';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
 import { io } from 'socket.io-client';
-import Api from '../services/index.js';
+import Api from '../../index.js';
 
 import { Box } from '@twilio-paste/core/box';
 import { Heading } from '@twilio-paste/core/heading';
