+++ apps/client/src/features/tasks/services/reports.js
@@ -1,3 +1,3 @@
-import http from './http.js';
+import http from '../../../shared/services/http.js';
 
 export const reports = () => http.get('/reports').then((r) => r.data);
