+++ apps/client/src/shared/components/CardSection.jsx
